+++ backend/src/main/java/com/pobluesky/backend/domain/file/dto/SimpleMultipartFile.java
@@ -1,40 +1,22 @@
-package com.pobluesky.backend.domain.file;
+package com.pobluesky.backend.domain.file.dto;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 import org.springframework.web.multipart.MultipartFile;
 
+@Getter
+@AllArgsConstructor
 public class SimpleMultipartFile implements MultipartFile {
 
     private final String name;
     private final String originalFilename;
     private final String contentType;
     private final byte[] content;
-
-    public SimpleMultipartFile(String name, String originalFilename, String contentType, byte[] content) {
-        this.name = name;
-        this.originalFilename = originalFilename;
-        this.contentType = contentType;
-        this.content = content;
-    }
-
-    @Override
-    public String getName() {
-        return this.name;
-    }
-
-    @Override
-    public String getOriginalFilename() {
-        return this.originalFilename;
-    }
-
-    @Override
-    public String getContentType() {
-        return this.contentType;
-    }
 
     @Override
     public boolean isEmpty() {
